split functions for host, players, spectators into different files, created new object for http socket io server, made the server io instance an import
